+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change trailer and combine quick form log categories. [#438](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/438)
+- Rename experiment code to study period ID. [#456](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/456)
 
 ### Fixed